+++ drools-core/src/main/java/org/drools/core/impl/KnowledgeBaseImpl.java
@@ -681,7 +681,7 @@ public class KnowledgeBaseImpl
         return this.factHandleFactory.newInstance();
     }
 
-    public FactHandleFactory newFactHandleFactory(int id,
+    public FactHandleFactory newFactHandleFactory(long id,
                                                   long counter) {
         return this.factHandleFactory.newInstance(id,
                                                   counter);
