+++ drools-core/src/main/java/org/drools/core/reteoo/BaseLeftTuple.java
@@ -16,12 +16,12 @@
 
 package org.drools.core.reteoo;
 
+import java.util.Arrays;
+
 import org.drools.core.common.InternalFactHandle;
 import org.drools.core.common.InternalWorkingMemory;
 import org.drools.core.spi.PropagationContext;
 import org.drools.core.util.index.TupleList;
-
-import java.util.Arrays;
 
 /**
  * A parent class for all specific LeftTuple specializations
@@ -599,7 +599,7 @@ public class BaseLeftTuple extends BaseTuple implements LeftTuple {
     protected String toExternalString() {
         StringBuilder builder = new StringBuilder();
         builder.append( String.format( "%08X", System.identityHashCode( this ) ) ).append( ":" );
-        int[] ids = new int[this.index+1];
+        long[] ids = new long[this.index+1];
         LeftTuple entry = this;
         while( entry != null ) {
             ids[entry.getIndex()] = entry.getFactHandle().getId();
